Move iceberg procedures and tests to package `...iceberg.procedure`
